+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Breaking changes
 * The `QuantumDevice.cfg_nr_averages` parameter was renamed to `QuantumDevice.cfg_sched_repetitions`
 * The call signature of `gettables.ScheduleVectorAcqGettable` has been renamed to `gettables.ScheduleGettableSingleChannel`, and the call signature has been updated according to #36 to no longer accept several keyword arguments.
 * The `RampPulse` has an extra (optional) parameter `offset` (!211)
+* Qblox Backend - The NCO phase is now reset at the start of a program (!213).
 
 Merged branches and closed issues
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~